HAR-65: support node removal after node is offline
In p2p library, I added a channel notification mechansim when it failed
to send messages after retries. The failed p2p.Peer will be notified to
caller via the channel. Then the leader knows which peer is offline and
adds to an offline list.

Before another consensus is started, the ledaer will check the offline
list and remove the node from the validator list.  Leader will have to
broadcast the Pong message to all validators to sync the public key
list.

Signed-off-by: Leo Chen <leo@harmony.one>