Check the output status of gogenerate.sh
This is because it may fail if the codegen tools are not found.

If failed, skip the status compare step because it may as well return a
false positive, i.e. generated code is not up to date but the codegen
tool to bring it up to date was not found so no diff was detected.