Resolve harmony-one/bounties#90: Add revert mechanism for UpdateValidatorWrapper (#3939)
* Add revert mechanism for UpdateValidatorWrapper

Closes harmony-one/bounties#90
(1) Use LRU for ValidatorWrapper objects in stateDB to plug a potential
    memory leak
(2) Merge ValidatorWrapper and ValidatorWrapperCopy to let callers ask
    for either a copy, or a pointer to the cached object. Additionally,
    give callers the option to not deep copy delegations (which is a
    heavy process). Copies need to be explicitly committed (and thus
    can be reverted), while the pointers are committed when Finalise
    is called.
(3) Add a UpdateValidatorWrapperWithRevert function, which is used by
    staking txs `Delegate`, `Undelegate`, and `CollectRewards`. Other
    2 types of staking txs and `db.Finalize` continue to use
    UpdateValidateWrapper without revert, again, to save memoery
(4) Add unit tests which check
    a) Revert goes through
    b) Wrapper is as expected after revert
    c) State is as expected after revert

* Change back to dictionary for stateValidators

Since the memory / CPU usage saved is not significantly different when
using an LRU + map structure, go back to the original dictionary
structure to keep code easy to read and have limited modifications.

* Add tests for validator wrapper reverts

As requested by @rlan35, add tests beyond just adding and reverting a
delegation. The tests are successive in the sense that we do multiple
modifications to the wrapper, save a snapshot before each modification
and revert to each of them to confirm everything works well. This change
improves test coverage of statedb.go to 66.7% from 64.8% and that of
core/state to 71.9% from 70.8%, and covers all the code that has been
modified by this PR in statedb.go.

For clarity, the modifications to the wrapper include (1) creation of
wrapper in state, (2) adding a delegation to the wrapper, (3)
increasing the blocks signed, and (4) a change in the validator Name and
the BlockReward. Two additional tests have been added to cover the
`panic` and the `GetCode` cases.