fix: make localnet ports consistent (#4270)
Since the introduction of port 9501 for RPC Auth HTTP Port, the default
localnet config produces a conflict in ports. In other words, the HTTP
Auth Port of the node from line 1 clashes with the HTTP Port of the node
from line 2. Depending on which of the two processes starts first, the
ports chosen by `make debug` can be different for each run, producing
inconsistencies in RPC test environments.