rpc: fix the `from` address calculation (#4593)
* test: add delegation type tests

These were skipped intentionally in the previous pull request

* rpc: undo all changes to 2023.3.0

* rpc: calculate SenderAddress before `ConvertToEth`

`tx.ConvertToEth` silently drops the `ShardID` and `ToShardID` fields.
This results in the hash of the transaction changing (either via removal
of these fields in the hash calculation or by automatic filling of the
values by the node's shard). The different hash calculation results in
incorrect sender address calculation.

There have been a couple of attempts to fix this issue, which created
troubles elsewhere. This pull request reverts to the behaviour seen in
2023.3.0 with a simple edit that calculates the `SenderAddress` before
dropping the fields in the call to `ConvertToEth`.